+++ packages/sdks/snippets/angular/src/app/announcement-bar/announcement-bar.component.ts
@@ -1,0 +1,56 @@
+/**
+ * https://www.builder.io/c/docs/integrate-section-building
+ * https://www.builder.io/c/blueprints/announcement-bar
+ * src/app/announcement-bar/announcement-bar.component.ts
+ */
+
+// fails because type imports cannot be injected
+// eslint-disable-next-line @typescript-eslint/consistent-type-imports
+import { ChangeDetectorRef, Component } from '@angular/core';
+import { fetchOneEntry, type BuilderContent } from '@builder.io/sdk-angular';
+
+@Component({
+  selector: 'app-announcement-bar',
+  template: `
+    <ng-container *ngIf="content; else notFound">
+      <content-variants
+        [model]="model"
+        [content]="content"
+        [apiKey]="apiKey"
+      ></content-variants>
+    </ng-container>
+
+    <ng-template #notFound>
+      <div>Announcement Bar not Found</div>
+    </ng-template>
+
+    <!-- Your content coming from your app (or also Builder) -->
+    <div>The rest of your page goes here</div>
+  `,
+})
+export class AnnouncementBarComponent {
+  apiKey = 'ee9f13b4981e489a9a1209887695ef2b';
+  model = 'announcement-bar';
+  content: BuilderContent | null = null;
+
+  constructor(private cdr: ChangeDetectorRef) {}
+
+  async ngOnInit() {
+    const urlPath = window.location.pathname || '';
+
+    const content = await fetchOneEntry({
+      apiKey: this.apiKey,
+      model: this.model,
+      userAttributes: {
+        urlPath,
+      },
+    });
+
+    if (!content) {
+      return;
+    }
+
+    this.content = content;
+    this.cdr.detectChanges();
+  }
+}
